Renaming core application directories and variables to MMCQ; removing external BL_VERSION, using default Biolink Model Toolkit() version instead
