+++ .env-template
@@ -3,12 +3,29 @@
 # whenever you change the contents of your '.env' copy of this file,  make sure that your
 # .env file is 'unix' style, with simple \n EOL characters (no Windoze \r carriage returns!)
 # using a bash tool like 'dos2unix' to make all line endings *nix compatible.
-WEB_HOST="127.0.0.1"
+
+# Monarch SemSimian BackEnd API Server
+
+# (Low Traffic) development environment parameters
+SEMSIMIAN_PROTOCOL="http://"
+SEMSIMIAN_HOST="api-v3.monarchinitiative.org"
+SEMSIMIAN_PORT=""  # default is HTTP port '80'
+
+# Alternate ("Docker Swarm") production environment parameters
+# SEMSIMIAN_PROTOCOL="http://"
+# SEMSIMIAN_HOST="api"
+# SEMSIMIAN_PORT="8000"
+
+SEMSIMIAN_SEARCH="/v3/api/semsim/search"
+
+# Web Host Deployment Parameters
+WEB_HOST="0.0.0.0"
 WEB_PORT="8080"
-MTA_SERVICE_ADDRESS="54.87.193.222"
-MTA_TITLE="Monarch TRAPI KP"
-MTA_VERSION="1.4.0"
-BL_VERSION="4.1.0"
+
+# TRAPI Wrapper
+MMCQ_SERVICE_ADDRESS="localhost"
+MMCQ_TITLE="Monarch SemSimian TRAPI Multi-CURIE Query Server"
+MMCQ_VERSION="1.4.0"
 #
 # Uncomment MODE if you wish to deploy
 # the application to gunicorn instead of uvicorn
