+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+# 0.0.2
+
+- June 2024 update to January prototype.
+
 # 0.0.1
 
 - Initial January 2024 prototype of the TRAPI phenotype-disease multi-curie query against the Monarch SemSimian.
