+++ docker-compose.yaml
@@ -4,12 +4,15 @@ services:
     build:
       context: .
     environment:
+      SEMSIMIAN_PROTOCOL: "${SEMSIMIAN_PROTOCOL}"
+      SEMSIMIAN_HOST: "${SEMSIMIAN_HOST}"
+      SEMSIMIAN_PORT: "${SEMSIMIAN_PORT}"
+      SEMSIMIAN_SEARCH: "${SEMSIMIAN_SEARCH}"
       WEB_HOST: "${WEB_HOST}"
       WEB_PORT: "${WEB_PORT}"
-      MTA_SERVICE_ADDRESS: "${MTA_SERVICE_ADDRESS}"
-      MTA_TITLE: "${MTA_TITLE}"
-      MTA_VERSION: "${MTA_VERSION}"
-      BL_VERSION: "${BL_VERSION}"
+      MMCQ_SERVICE_ADDRESS: "${MMCQ_SERVICE_ADDRESS}"
+      MMCQ_TITLE: "${MMCQ_TITLE}"
+      MMCQ_VERSION: "${MMCQ_VERSION}"
       MODE: "${MODE}"
       OTEL_ENABLED: "${OTEL_ENABLED}"
       OTEL_USE_CONSOLE_EXPORTER: "${OTEL_USE_CONSOLE_EXPORTER}"
