+++ mmcq/services/app_common.py
@@ -22,11 +22,11 @@ from fastapi import (
 #     PredicatesResponse
 # )
 # from mtkp.services.util.bl_helper import BLHelper
-from mtkp.services.util.monarch_adapter import MonarchInterface
-from mtkp.services.util.metadata import GraphMetadata
+from mmcq.services.util.monarch_adapter import MonarchInterface
+from mmcq.services.util.metadata import GraphMetadata
 # from mtkp.services.util.overlay import Overlay
 # from mtkp.services.util.question import Question
-from mtkp.services.util.api_utils import (
+from mmcq.services.util.api_utils import (
     get_monarch_interface,
     # get_bl_helper,
     construct_open_api_schema,
