+++ mmcq/services/server.py
@@ -5,15 +5,15 @@ import os
 
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
-from mtkp.services.config import config
-from mtkp.services.util.logutil import LoggingUtil
-from mtkp.services.app_common import APP_COMMON
-from mtkp.services.app_trapi_1_4 import APP_TRAPI_1_4
-from mtkp.services.util.api_utils import construct_open_api_schema
+from mmcq.services.config import config
+from mmcq.services.util.logutil import LoggingUtil
+from mmcq.services.app_common import APP_COMMON
+from mmcq.services.app_trapi_1_4 import APP_TRAPI_1_4
+from mmcq.services.util.api_utils import construct_open_api_schema
 
-TITLE = config.get('MTA_TITLE', 'Monarch TRAPI KP')
+TITLE = config.get('MMCQ_TITLE', 'Monarch TRAPI KP')
 
-VERSION = os.environ.get('MTA_VERSION', '1.4.0')
+VERSION = os.environ.get('MMCQ_VERSION', '1.4.0')
 
 logger = LoggingUtil.init_logging(
     __name__,
